refactors development branch creation
