+++ src/commands/createDevelopmentBranch.js
@@ -1,0 +1,7 @@
+import run from '../utils/run';
+
+export default () => run([
+  'cd tmp/repo',
+  'git checkout -b development',
+  'git push -u origin development',
+]);
