+++ src/commands/deleteRepo.js
@@ -1,0 +1,5 @@
+import run from '../utils/run';
+
+export default () => run([
+  'rm -rf tmp/repo',
+]);
